fastapi/requirements.txt
fastapi
uvicorn
streamlit
google-generativeai
requests